Move SH*TE declaration to its own file
But also where all other events should be
declared, health/events.go
