+++ health/events.go
@@ -1,0 +1,32 @@
+package health
+
+// Events
+
+type TranscodeEvent struct {
+	NodeID     string                 `json:"nodeId"`
+	ManifestID string                 `json:"manifestId"`
+	Segment    SegmentMetadata        `json:"segment"`
+	StartTime  int64                  `json:"startTime"`
+	LatencyMs  int64                  `json:"latencyMs"`
+	Success    bool                   `json:"success"`
+	Attempts   []TranscodeAttemptInfo `json:"attempts"`
+}
+
+// TranscodeEvent inner types
+
+type OrchestratorMetadata struct {
+	Address       string `json:"address"`
+	TranscoderUri string `json:"transcodeUri"`
+}
+
+type TranscodeAttemptInfo struct {
+	Orchestrator OrchestratorMetadata `json:"orchestrator"`
+	LatencyMs    int64                `json:"latencyMs"`
+	Error        *string              `json:"error"`
+}
+
+type SegmentMetadata struct {
+	Name     string  `json:"name"`
+	SeqNo    uint64  `json:"seqNo"`
+	Duration float64 `json:"duration"`
+}
